Add skellam.cc to avoid double definitions of lognormal_logp.
