+++ cxx/relation_test.cc
@@ -48,7 +48,7 @@ BOOST_AUTO_TEST_CASE(test_relation) {
   lpg = R1.logp_gibbs_approx(D1, 0, 10, &prng);
   R1.set_cluster_assignment_gibbs(D1, 0, 1, &prng);
 
-  Distribution<bool>* db = R1.make_new_distribution();
+  Distribution<bool>* db = R1.make_new_distribution(&prng);
   BOOST_TEST(db->N == 0);
   db->incorporate(false);
   BOOST_TEST(db->N == 1);
@@ -64,7 +64,7 @@ BOOST_AUTO_TEST_CASE(test_relation) {
   R2.set_cluster_assignment_gibbs(D3, 3, 1, &prng);
   D1.set_cluster_assignment_gibbs(0, 1);
 
-  Distribution<std::string>* db2 = R2.make_new_distribution();
+  Distribution<std::string>* db2 = R2.make_new_distribution(&prng);
   BOOST_TEST(db2->N == 0);
   db2->incorporate("hello");
   BOOST_TEST(db2->N == 1);
